+++ src/client.ts
@@ -4,8 +4,9 @@ import type { AxiosAdapter, CreateAxiosDefaults, AxiosInstance, AxiosProxyConfig
 import { SdkError, handleError } from './error'
 import type { InterceptorManager } from './interceptor'
 import config from './config'
-import type { Agent as HttpAgent } from 'http'
-import type { Agent as HttpsAgent } from 'https'
+// import type { Agent as HttpAgent } from 'http'
+// import type { Agent as HttpsAgent } from 'https'
+// import { packageInfo } from './util'
 
 import Debug from './debug'
 const debug = Debug('client')
@@ -25,21 +26,26 @@ type RequestHeaders = Record<string, string>
 
 // Subset of AxiosRequestConfig
 type RequestConfig = {
-	timeout?: number;
-	params?: RequestParams;
-	httpAgent?: HttpAgent;
-	httpsAgent?: HttpsAgent;
-	proxy?: ProxyConfig;
+	timeout?: number
+	params?: RequestParams
+	httpAgent?: any
+	httpsAgent?: any
+	proxy?: ProxyConfig
 	headers?: RequestHeaders
 }
 
+type RequestConfigExtra = {
+	adapter?: Adapter
+	userAgent?: string
+}
+
 type ApiConfig = {
-	organization: string,
-	domain?: string,
+	organization: string
+	domain?: string
 	accessToken: string
 }
 
-type ApiClientInitConfig = ApiConfig & RequestConfig & { adapter?: Adapter }
+type ApiClientInitConfig = ApiConfig & RequestConfig & RequestConfigExtra
 type ApiClientConfig = Partial<ApiClientInitConfig>
 
 
@@ -68,11 +74,16 @@ class ApiClient {
 			timeout: options.timeout || config.client.timeout,
 			proxy: options.proxy,
 			httpAgent: options.httpAgent,
-			httpsAgent: options.httpsAgent,
+			httpsAgent: options.httpsAgent
 		}
 
 		// Set custom headers
 		const customHeaders = this.customHeaders(options.headers)
+
+		// Set User-Agent
+		// const userAgentData = packageInfo(['version', 'dependencies.axios'], { nestedName: true })
+		let userAgent = options.userAgent || `SDK-core axios/${axios.VERSION}`
+		if (!userAgent.includes('axios/')) userAgent += ` axios/${axios.VERSION}`
 
 		const axiosOptions: CreateAxiosDefaults = {
 			baseURL: this.baseUrl,
@@ -81,7 +92,8 @@ class ApiClient {
 				...customHeaders,
 				'Accept': 'application/vnd.api+json',
 				'Content-Type': 'application/vnd.api+json',
-				'Authorization': 'Bearer ' + this.#accessToken
+				'Authorization': 'Bearer ' + this.#accessToken,
+				'User-Agent': userAgent
 			},
 			...axiosConfig
 		}
@@ -109,6 +121,10 @@ class ApiClient {
 		if (config.httpAgent) def.httpAgent = config.httpAgent
 		if (config.httpsAgent) def.httpsAgent = config.httpsAgent
 
+		if (config.adapter) this.adapter(config.adapter)
+		if (config.userAgent) this.userAgent(config.userAgent)
+
+
 		// API Client config
 		if (config.organization) this.baseUrl = baseURL(config.organization, config.domain)
 		if (config.accessToken) {
@@ -116,10 +132,22 @@ class ApiClient {
 			def.headers.common.Authorization = 'Bearer ' + this.#accessToken;
 		}
 		if (config.headers) def.headers.common = this.customHeaders(config.headers)
-		if (config.adapter) this.adapter(config.adapter)
 
 		return this
 
+	}
+
+
+	userAgent(userAgent: string): ApiClient {
+		if (userAgent) {
+			let ua = userAgent
+			if (!ua.includes('axios/')) {
+				// const axiosVer = packageInfo(['dependencies.axios'], { nestedName: true })
+				if (axios.VERSION) ua += ` axios/${axios.VERSION}`
+			}
+			this.#client.defaults.headers['User-Agent'] = ua
+		}
+		return this
 	}
 
 
@@ -133,11 +161,15 @@ class ApiClient {
 
 		debug('request %s %s, %O, %O', method, path, body || {}, options || {})
 
+		// Ignored params alerts (in debug mode)
+		if (options?.adapter) debug('Adapter ignored in request config')
+		if (options?.userAgent) debug('User-Agent header ignored in request config')
+
 		const data = body ? { data: body } : undefined
 		const url = path
 
 		// Runtime request parameters
-		const baseUrl = options?.organization ? baseURL(options.organization, options.domain) : undefined
+		const baseUrl = options?.organization? baseURL(options.organization, options.domain) : undefined
 		const accessToken = options?.accessToken || this.#accessToken
 
 		const headers = this.customHeaders(options?.headers)
@@ -160,7 +192,7 @@ class ApiClient {
 		const customHeaders: RequestHeaders = {}
 		if (headers) {
 			for (const [name, value] of Object.entries(headers))
-				if (!['accept', 'content-type', 'authorization'].includes(name.toLowerCase())) customHeaders[name] = value
+				if (!['accept', 'content-type', 'authorization', 'user-agent'].includes(name.toLowerCase())) customHeaders[name] = value
 		}
 		return customHeaders
 	}