+++ src/util.ts
@@ -1,4 +1,5 @@
-import type { ObjectType } from "../src/common";
+import type { ObjectType } from "../src/common"
+// import path from 'node:path'
 
 
 
@@ -15,5 +16,32 @@ const sortObjectFields = (obj: ObjectType): ObjectType => {
 	return sorted
 }
 
+/*
+const nestedField = (obj: any, field: string): { key: string, val: any } => {
 
-export { sleep, sortObjectFields }
+	let fp = field
+	if (fp.endsWith('.')) fp = fp.substring(0, fp.length-1)
+
+    const dots = field.split(".")
+
+	const key = dots[dots.length-1]
+	let val = obj
+    while (dots.length && (val = val[dots.shift() || '']));
+    
+	return { key, val }
+}
+*/
+
+/*
+const packageInfo = (fields?: string | string[], options?: any): Record<string, any>  => {
+	const pjson = require(path.resolve('./', 'package.json'))
+	return fields? (Array.isArray(fields)? fields : [ fields ]).reduce((info: any, field) => {
+		const nf = nestedField(pjson, field)
+		info[options?.nestedName? nf.key : field] = nf.val
+		return info
+	  }, {}) : pjson
+}
+*/
+
+
+export { sleep, sortObjectFields, /* packageInfo */ }